move generator to new folder
