+++ nn_meter/builder/model_generator/generator/networks/block_utils.py
