+++ nn_meter/builder/rule_tester/__init__.py
@@ -1,3 +1,3 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
-from .rule_tester import RuleTester
+from .runner import RuleTester
