--- nn_meter/builder/rule_tester/rule_tester/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-from .runner import RuleTester
